backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pytz>=2023.3
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
requests>=2.31.0
slowapi>=0.1.9
sqlalchemy>=2.0.0
asyncpg>=0.29.0
